requirements.txt
PyQt6>=6.6.0
PyOpenGL>=3.1.7
PyOpenGL_accelerate>=3.1.7
numpy>=1.26.0
plyfile>=1.0.3
pycpd>=2.0.0
